More gcode engine decomposition
