+++ docs/refactor-architecture-plan.md
@@ -641,6 +641,7 @@ Use this section to track completed steps with date and PR/commit references.
 - 2026-04-10: Continued gcodeEngine pipeline modularization with stage extraction of coordinate transform + curve flattening: moved `transformPt`, `cubicBezier`, `quadBezier`, and `arcToBeziers` into `src/workers/gcodeEngine/stages/geometryFlattening.ts`, kept backward-compatible re-exports in `src/workers/gcodeEngine.ts`, and added focused direct stage coverage in `tests/unit/gcodeGeometryFlattening.test.ts`.
 - 2026-04-10: Continued gcodeEngine pipeline modularization with path optimization stage extraction: moved `nearestNeighbourSort` + `joinSubpaths` into `src/workers/gcodeEngine/stages/pathOptimization.ts`, kept backward-compatible re-exports in `src/workers/gcodeEngine.ts`, and added focused deterministic stage coverage in `tests/unit/gcodePathOptimization.test.ts`.
 - 2026-04-10: Continued gcodeEngine pipeline modularization with clipping stage extraction: moved `clipSubpathsToRect` + `clipSubpathsToBed` (and internal Liang-Barsky clipping logic) into `src/workers/gcodeEngine/stages/clipping.ts`, kept backward-compatible re-exports in `src/workers/gcodeEngine.ts`, and added focused deterministic stage coverage in `tests/unit/gcodeClipping.test.ts`.
+- 2026-04-10: Continued gcodeEngine pipeline modularization with flattening-flow stage extraction: moved `flattenToSubpaths` into `src/workers/gcodeEngine/stages/flatteningFlow.ts`, preserved its use of the extracted parsing + geometry stages, kept backward-compatible re-exports in `src/workers/gcodeEngine.ts`, and added focused deterministic stage coverage in `tests/unit/gcodeFlatteningFlow.test.ts`.
 - 2026-04-10: Started gcodeEngine pipeline modularization with stage extraction of path parsing: moved `tokenizePath` + `toAbsolute` (and `PathToken`) into `src/workers/gcodeEngine/stages/pathParsing.ts`, kept backward-compatible exports in `src/workers/gcodeEngine.ts`, and added focused deterministic coverage in `tests/unit/gcodePathParsing.test.ts`.
 
 ## Update Rule
