+++ src/workers/gcodeEngine.ts
@@ -26,6 +26,7 @@ import {
   clipSubpathsToRect,
   clipSubpathsToBed,
 } from "./gcodeEngine/stages/clipping";
+import { flattenToSubpaths } from "./gcodeEngine/stages/flatteningFlow";
 
 // ── Types ─────────────────────────────────────────────────────────────────────
 
@@ -68,150 +69,7 @@ export { nearestNeighbourSort };
 export { joinSubpaths };
 
 // ── Main flattener ────────────────────────────────────────────────────────────
-
-export function flattenToSubpaths(
-  obj: VectorObject,
-  config: MachineConfig,
-): Subpath[] {
-  const subpaths: Subpath[] = [];
-  const abs = toAbsolute(tokenizePath(obj.path));
-  let cur: Subpath = [];
-  let cx = 0,
-    cy = 0,
-    startX = 0,
-    startY = 0;
-  let lastCpX = 0,
-    lastCpY = 0;
-
-  const push = (x: number, y: number) => {
-    cur.push(transformPt(obj, config, x, y));
-    cx = x;
-    cy = y;
-  };
-
-  for (const cmd of abs) {
-    const t = cmd.type;
-    const a = cmd.args;
-    switch (t) {
-      case "M": {
-        if (cur.length > 1) subpaths.push(cur);
-        cur = [];
-        cx = a[0];
-        cy = a[1];
-        startX = cx;
-        startY = cy;
-        cur.push(transformPt(obj, config, cx, cy));
-        lastCpX = cx;
-        lastCpY = cy;
-        for (let i = 2; i < a.length; i += 2) push(a[i], a[i + 1]);
-        break;
-      }
-      case "L":
-        for (let i = 0; i < a.length; i += 2) push(a[i], a[i + 1]);
-        lastCpX = cx;
-        lastCpY = cy;
-        break;
-      case "H":
-        for (let i = 0; i < a.length; i++) push(a[i], cy);
-        lastCpX = cx;
-        lastCpY = cy;
-        break;
-      case "V":
-        for (let i = 0; i < a.length; i++) push(cx, a[i]);
-        lastCpX = cx;
-        lastCpY = cy;
-        break;
-      case "Z":
-        push(startX, startY);
-        subpaths.push(cur);
-        cur = [];
-        lastCpX = cx;
-        lastCpY = cy;
-        break;
-      case "C":
-        for (let i = 0; i < a.length; i += 6) {
-          for (const p of cubicBezier(
-            cx,
-            cy,
-            a[i],
-            a[i + 1],
-            a[i + 2],
-            a[i + 3],
-            a[i + 4],
-            a[i + 5],
-          ))
-            push(p.x, p.y);
-          lastCpX = a[i + 2];
-          lastCpY = a[i + 3];
-        }
-        break;
-      case "S":
-        for (let i = 0; i < a.length; i += 4) {
-          const c1x = 2 * cx - lastCpX,
-            c1y = 2 * cy - lastCpY;
-          for (const p of cubicBezier(
-            cx,
-            cy,
-            c1x,
-            c1y,
-            a[i],
-            a[i + 1],
-            a[i + 2],
-            a[i + 3],
-          ))
-            push(p.x, p.y);
-          lastCpX = a[i];
-          lastCpY = a[i + 1];
-        }
-        break;
-      case "Q":
-        for (let i = 0; i < a.length; i += 4) {
-          for (const p of quadBezier(
-            cx,
-            cy,
-            a[i],
-            a[i + 1],
-            a[i + 2],
-            a[i + 3],
-          ))
-            push(p.x, p.y);
-          lastCpX = a[i];
-          lastCpY = a[i + 1];
-        }
-        break;
-      case "T":
-        for (let i = 0; i < a.length; i += 2) {
-          const cpx = 2 * cx - lastCpX,
-            cpy = 2 * cy - lastCpY;
-          for (const p of quadBezier(cx, cy, cpx, cpy, a[i], a[i + 1]))
-            push(p.x, p.y);
-          lastCpX = cpx;
-          lastCpY = cpy;
-        }
-        break;
-      case "A":
-        for (let i = 0; i < a.length; i += 7) {
-          for (const p of arcToBeziers(
-            cx,
-            cy,
-            a[i],
-            a[i + 1],
-            a[i + 2],
-            a[i + 3],
-            a[i + 4],
-            a[i + 5],
-            a[i + 6],
-          ))
-            push(p.x, p.y);
-          lastCpX = cx;
-          lastCpY = cy;
-        }
-        break;
-    }
-  }
-  if (cur.length > 1) subpaths.push(cur);
-  return subpaths;
-}
+export { flattenToSubpaths };
 
 // ── Bed clipping ─────────────────────────────────────────────────────────────
